02 - Local API and Interceptors
Class 2 completed successfully? So, let's remember what we learned:

Configuring the Local API: We have seen that there are variables that involve the configuration and maintenance of an API running on the Web, and that, to avoid inconvenience, we decided to use a local API for our studies. This API is the JSON-Server of Node.JS.

Services in Flutter: We also learned that it is important to separate the code blocks responsible for communicating our application with the APIs. To do this, we created a class called JournalService that will manage all our requests.

IP and HTTP: Next we understand a little of how the Internet works and what the IP and HTTP protocols are. We have seen that IP is a protocol that helps to locate a machine on the Web through an address and a port. HTTP, on the other hand, is the protocol we use to make client-server communications in order to access and manipulate the data that the server has.

Interceptors and Loggers: Finally, we discussed that, as the http library does not give us an interesting way to make a real-time diagnosis of the requests and responses that are going through our application, it would be important to configure one. And that's exactly what we did using the http_interceptor library to intercept all requests and responses; and the logger to show useful information in an organized and pleasant way on the console.