03 - API Registration
Another class finished and how about we review what we learned?

Asynchronousism: We started by doing a brief review of asynchronousism in Dart, because we noticed that when we are dealing with requests for Web APIs, these operations will not be instantaneous, that is, they will take a while to happen, so they will be asynchronous.

Input Addition Screen: We also saw that it was necessary to build a screen responsible for capturing from the user the information that will be sent to the Web API.

POST method: Next, we use the POST method to send the captured information to our Web API. To do this, it was necessary to convert our object to a JSON, configure the header to warn you that we are sending the information using this pattern and verify that the server response indicates that the entry was successfully created.

SnackBar: Finally, we noticed that it is extremely important to give feedback to the user about the success of the asynchronous operation. For this, we use a SnackBar that will be shown on the screen when the server response is a 201 Created.