04 - Listing on Initial Screen
Uhuu!! We arrived at our last class of the course. Let's remember what we learned?

Searching from the API: After learning how to add new information to a resource using POST, it was time for us to learn how to read the resource entries. For this we use the GET method to access the journals of our server and thus receive a list containing all registered entries.

Listing on the Screen: With the list of entries available via our operation with GET, we just needed to be able to show on the main screen. For this, we converted back the JSON that arrives as an answer to a list of Journals, and with this list we can replace the random database we used with this real information received via Web API.

Refactoring the code: Finally, we refactored our code so that it was much more structured and readable.